Fixes a static initialization bug under linux. (#571)
The osGlobalAtomic variable was accessed from the non-default threads while the
static initializers were running. This caused the mutex to be lost sometimes.
After this fix a static Atomic variable does not have a constructor call, but is
linker-initialized from the .data section.

This bug caused the hub application to get sometimes "livelocked" on linux, with the
packets printed to the screen but not being sent to any connected devices actually.