Adds platform-independent interface for sleeping. (#577)
The intention is to replace all usage of deprecated POSIX function `usleep` with the function defined in this cross-platform header. Then different platforms can provide their implementation.

This will allow us to remove the feature macro hacks that we had to put in to make the code compile on recent releases of armgcc, such as `#define _BSD_SOURCE` or `#define _DEFAULT_SOURCE`. These are harmful, because they often conflict with each other and sometimes with expectations from middleware libraries (e.g. a macro declaring C99 source).

===

* Adds platform-independent interface for sleeping.

* Fix typo