Adds support for distcc (#593)
Updates the compilation scripts to be able to wrap compilations with distcc.
Adds documentation on howto set up distcc. Adds helper script to start correct SSH tunnel. Adds auto-detection on whether the tunnel is up or not.

This dramatically reduces compilation time for complex projects, sometimes by a factor of 2x-3x if a single powerful backend machine is used.

Fixes all usages of .d (dependency) file generation to skip the system headers.

===

* Ensures that we always use -MMD -MF to generate dependency files.
This will remove system headers from the dependencies, which is important
because the system headers might change between compiler versions.

* Adds distcc scripts and howto documentation.

* Updates cortex-m targets to use distcc when available.

* Adds more comments to the docs.

* Fixes formatting in documentation.

* Fix some style in the doc.

* Kills the remote distccd before trying to start it.

* Enables forwarding g++ and gcc invocations as well.

* Enables distcc for test compilations.