Explicit handling of NOT_RESPONDING aliases.
Ensures that we can enter more than one ID with NOT_RESPONDING alias into the alias cache.
This is critical so that the AddressedWriteFlow can correctly function in the presence
of node IDs that disappeared from the network (or never existed).

We already have a workaround where we add a node ID with the special alias NOT_RESPONDING,
but the AliasCache was not prepared for having more than one such entry, and thus it was
evicting these entries unexpectedly.

This PR makes the NOT_RESPONDING entries unique in the aliasMap by appending part of the
storage offset to the alias.