Adds a virtual function to If object called `get_default_node_id`.
This function returns the node ID of the default node on an interface. For single-node interfaces
like SimpleStack this is always the same as the node ID of that single node.

Generally SimpleStack takes a NodeID on the constructor, and that node ID
is returned.

This is well defined, because a default node ID is needed for both CAN and TCP
interfaces by the transport standard:
- The CAN interface needs a node ID while allocating aliases.
- The TCP interface needs to add a last gateway node ID into every message header.