Optimize latency of hub (#667)
Adds a unit test which exercises a gridconnect hub and a client connecting through a socket exchanging a long sequence of datagrams. The total time is measured for how long it takes to ping-pong datagrams back and forth. This is a decent simulation of what is happening when a lot of data (e.g. CDI) is downloaded using memory config protocol.

Adds an optimization to the hub implementation that flushes the data from the buffer when certain latency-critical messages are recognized. The latency critical messages are always the last messages from one node after which a response from the other node is expected: 1) datagram messages (only the last frame) and 2) datagram received OK messages.

===

* Adds a test that verifies the performance (latency) of datagram ping-pong protocols.

Optimizes the latency internal to the gridconnecthub using an openlcb-specific hardocded logic.

* switches to tcp server+socket.

* Adds comment.