Respond Ack to memcfg write datagrams immediately.
Fixes a TODO where the memory config write datagram was acknowledged only after the write has succeeded.
This created an extra delay that could cause a timeout on the caller.
Waiting for the write to complete was not necessary, as success/error return
status goes in a write response datagram anyway.