Adds time jump information to the broadcast time update callbacks. (#694)
* Adds time jump information to the broadcast time update callbacks.

The jump information gives direct access in the callback to both the old state
and the new state. This allows the user of the time object to determine whether
the fast time has jumped, by how much and which direction.

Use-cases that work with time intervals need this jump information.