Adds 9bit reception support to the TI uart drivers. (#713)
Implements a 9-bit reception mode for Tiva and CC32xx UART drivers.
The nine-bit reception works by reading two byte payloads from the fd for each incoming byte, LSB-first. The second byte contains the parity bit. The special mode is enabled with an ioctl, and only works for RX, not for TX.

===

* Adds 9bit reception capability to the CC32xxUart.

* Adds 9bit to the tiva UART as well.

* Fix comments.