Limit the amount of input we read into memory in arduino sketches.
Adds a pushback mechanism to not read faster from a serial bridge
than how fast the stack, or a CAN-bus output can sink packets.
This is especially important for USBSerial inputs, which can supply
data as fast as we are willing to read it.

The implementation relies on two mechanisms:
- setting the gridconnect converter to keep track of the CAN frames and only let
  a fixed number of them (5) exist at any time in the CAN hub.
- adding a notifiable to the SerialBridge that keep track of the last txt buffer
  we sent to the gridconnect bridge. We do not read more bytes from the serial
  port in the read loop until the previous buffer is processed by the gridconnect parser.