Adds TinyUsb based CDC virtual Com port driver with an STM32 port. (#748)
- Adds TinyUsb to path.mk
- Adds a target directory to compile it for the stm32f072xb MCU
- Adds a common base class for a TinyUsb based USB stack with CDC device driver
- Instantiates this for STM32
- Adds a CAN-USB application target for the STM32F072xB.

The TinyUSBCdc driver is select-enabled. It does not use a Serial base class, because the read and write fifo's are provided by the TinyUsb codebase. We are exercising the application reads and writes directly from those fifos; this saves one unnecessary copy of the data in memory. (There are still two memcpy's; one in the interrupt to the fifo, one in the application code to the application buffer.)

===

* Adds a CAN-USB target for an stm32f072 board.

* Adds a sources + target directory for compiling tinyusb for the stm32F072 MCU.

* Adds tinyusb to path.mk

* Recurses into the tinyusb f072 subdirectory during build.

* Adds the common (MCU independent part) of the TinyUSB-CDC driver.

* Adds stm32 specific USB driver class.

* Minor compilation fixes.

* Implements buffer management and select wakeups in the tinyusb cdc driver.

* Adds common makefile to make it easier to write tinyusb per-MCU-target makefiles.

* Completes setup code.

* Fix missing flush.
OPtimize thread stack size.

* fix comments

* Fixes whitespace.

* Optimize the RAM needs for the CAN-USB application (#749)

The canusb application could not run in 16kbyte MCU's because it was running out of heap. This PR optimizes it, and handles some long standing work items in the buffering capabilities of openmrn.

- switches can-usb app to use select() based implementation.
- Removes five tasks:
  - CAN-bus read/write threads
  - Serial port read/write tasks
  - merges the main thread and the main executor thread
- Fixes traffic throttling on HubDeviceSelect. It was not on par with HubDevice, as the input data was consumed at an infinite rate. Adds support for the linker option config_gridconnect_port_max_incoming_packets().

===

* Add support for LimitedPool to HubDeviceSelect.
This brings it on par with HubDevice in ensuring that we don't allow too much
data into the memory of the embedded device, and push back on the traffic stream
on the source (USB or TCP should both work).
The limit will be based on the linker option config_gridconnect_port_max_incoming_packets().

* Switches the USB-CAN application to use select()-based implementation.
This dramatically reduces the RAM requirements.

Removes five tasks:
- CAN-bus read/write threads
- Serial port read/write tasks
- merges the main thread and the main executor thread