Limits end to end buffering in gridconnect bridge.
When a dispatcher was used to send traffic to a gridconnect bridge,
an infinite number of packets were allocated in the dispatcher when more than
one client was connected.

By adding a limitedpool to the bridge object, we can ensure that the dispatcher is blocked when
the output queue pushes back.