Adds support for Offset(n) attribute in the CDI  (#765)
Adds support for Offset(n) attribute in the CDI for individual entries, including groups, atoms (like string) or number entries.

It is also supported to declare a negative offset, which is a typical use- case when the entries are displayed out of order, or when one source data needs to be displayed in two ways.

Previous Offset() was only supported on segments, which are a form of groups. With this PR, the declaration is moved up to the root class of all config entries. From then on every entry must have some form of definition.
Also adjusts how the c++ constexpr generates the offset() attribute to take this into account.