Adds an application (hub_test) for testing the throughput of a hub or a CAN-bus (#762)
Adds a new application to OpenMRN called hub_test. This application can generate a specific amount of traffic, send it to a bus (a CAN-bus or a gridconnect-TCP hub). A second component is a receiver, which listens to traffic on the same OpenLCB bus. It verifies that data arrives without dropped packets, in order, and prints statistics about throughput and round-trip time (which correlates with buffer length).

===

* Adds the skeleton of a hub test application which will verify the performance
of a hub including packet loss.

* Adds an accessor to the last expiry time of a timer.

* Starts adding the send flow.

* Adds code for filling in the output frames.
Adds a structure for recording timestamps per packet.

* Releases packets as the receive notifications are coming in.
Adds stats printing timer.

* Adds receiver class to hub_test.

* hub-test: Implements links and prints statistics about sent data.

* Adds a utility class for computing simple statistics.

* fix whitespace

* hubtest: tune TCP options.

* Changes TCP tuning params: no snd/recv buf limitation, but notsent_lowat with 1460.

* Adds new features to the hub_test:
- support creating multiple TCP clients with -N number
- support sending different packets, like identify producer and verify node ID
  which allows measuring response latency of a node.

* Removes output buffer delay.