Adds support for DCC extended accessories  (#769)
- Adds helper functions to generate dcc extended accessory packets in dcc::Packet
- Adds well-known event range (from the current draft eventidentifiers standard)
- Adds DccExtAccyConsumer, which listens to events and turns them into DCC packets.

Unlike the basic dcc accy consumer, the extended consumer does not remember whatthe last set state was. This is because it takes a lot more memory (2 kbytes), and it's a lot less useful, since different decoders may be interpreting this last set command very differently (such as a signal, a turntable, or a timed turnout decoder).

===

* Adds command to generate dcc extended accessory packet.

* Adds well-known range for dcc extended accessories (according to the draft event identifiers standard).

* Adds accessory consumer for extended accessories.
Generates the dcc packets on the rails.