Fix compiler warnings in openmrn when using new GCC's. (#772)
There are a few places where a newer GCC used by esp-idf emits a warning for OpenMRN code. This PR fixes those.

The fall through warning unfortunately has false positives, but that's still not a good reason to disable it, we rather adjust the comments here to suit the taste of the compiler. 

===

* Fix compiler warnings in openmrn when using new GCC's.

* Fix comment.