Fixes a standards compliance issue with the alias conflict handler.
The standard requires that at the time when an alias conflict is detected,
the device to issue an AMR frame with the conflicted alias.

This PR adds the missing AMR frame. Adjusts tests to heck for this frame.
Fixes a bug in the Global AME handler that was unexpectedly emitting AMD
frames for removed aliases, because the cache entry was not correctly wiped
when the remove was processed, and the iteration also picked up such deleted
entries.