Fixes missing translation of enums when reading the security mode from a simplelink profile. (#796)
The external enum is not the same as the internal enum. We have translation functions `security_type_from_simplelink` and `security_type_to_simplelink`. We forgot to call this function on one of the codepaths.

Updates the setup_ap function to be able to be called with nullptr as ssid and security key, to just update the security type.

===

* Fixes missing translation of enums when reading the security mode from a
simplelink profile.

* Updates the setup AP API to be able to update just the security type.