Fixes the case when there is empty CONFIG_FILENAME in simplestack.  (#805)
We can build very simple nodes with no eeprom and set an empty CONFIG_FILENAME in simplestack. THere were several crashing bugs (regressions).

- The SNIP dynamic filename was not correctly taken into account.
- The SnipDynamicUserFile was not readable due to a regression in RamDisk.

===

* Fixes the case when there is empty CONFIG_FILENAME in simplestack. The SNIP dynamic filename
was not correctly taken into account.

* Makes a RamDisk initialized from a struct take this as the actual size
upon construction. This means if the struct has some backing data,
that will be visible in the file. Since #677 all read-write ramdisks
were initialized as empty, which is not the correct behavior when
they are created with pre-filled backing storage.

This fixes SNIP mock user file on freertos devices.