Fix mutex lock allocation during static initialization. (#868)
* Fix mutex lock allocation during static initialization.

This PR avoids the starting/stopping the scheduler when the scheduler
is not running yet. This is typically the case in code that is running
at static initialization time.

More recent freertos versions crash with an assert when we were doing that.

As a byside, makes this code more efficient to run by not even attempting
to manipulate the scheduler when the mutex is already allocated.
Since there is exactly one moment when sem goes from NULL to non-NULL, there
is no race condition if we observe it being non-NULL.

* Update variable name.