Uses the ESP API for disabling idle watchdog. (#869)
The Arduino API disableCore0WDT() for doing the same creates an infinite error loop
"wdt_reset(): task not found".
This is because it calls the API for deleting the watchdog for a specific
task handle, but does not unregister the idle task callback. So the idle task
keeps feeding the watchdog, causing the errors to be reported.